Optimizes repayment schedule with dp
